# src/main/resources/swagger.yaml
openapi: 3.0.3
info:
  title: Watch the Chamber - BFF API
  description: >
    Backend For Frontend API between the mobile app and the Brazilian Chamber of Deputies public API.
    Handles authentication, user management, followed politicians, and notifications.
  version: 1.0.0
  contact:
    name: Raphael Coelho
    email: contato@raphaelcoelho.dev
servers:
  - url: http://localhost:8080/api/v1
    description: Local environment
  - url: https://api.deolhonacamara.com/api/v1
    description: Production
...
paths:
...
  /api/v1/sync/all:
    post:
      tags: [ Sync ]
      summary: Trigger full synchronization of all features
      operationId: syncAll
      description: Starts an asynchronous synchronization process for all features (politicians, expenses, votes, speeches, propositions, and presence). Returns immediately without waiting for completion.
      responses:
        '202':
          description: Synchronization started successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncResponse'
        '500':
          description: Error starting synchronization
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SyncResponse:
      type: object
      properties:
        message:
          type: string
          example: Synchronization started successfully
        status:
          type: string
          example: ACCEPTED