# target/generated-sources/openapi/src/main/java/net/coelho/deolhonacamara/api/SyncApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.13.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package net.coelho.deolhonacamara.api;

import net.coelho.deolhonacamara.api.model.SyncResponse;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-11-21T10:53:37.437611200-03:00[America/Fortaleza]", comments = "Generator version: 7.13.0")
@Validated
@Tag(name = "Sync", description = "Data synchronization endpoints")
public interface SyncApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /api/v1/sync/all : Trigger full synchronization of all features
     * Starts an asynchronous synchronization process for all features (politicians, expenses, votes, speeches, propositions, and presence). Returns immediately without waiting for completion.
     *
     * @return Synchronization started successfully (status code 202)
     *         or Error starting synchronization (status code 500)
     */
    @Operation(
        operationId = "syncAll",
        summary = "Trigger full synchronization of all features",
        description = "Starts an asynchronous synchronization process for all features (politicians, expenses, votes, speeches, propositions, and presence). Returns immediately without waiting for completion.",
        tags = { "Sync" },
        responses = {
            @ApiResponse(responseCode = "202", description = "Synchronization started successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SyncResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Error starting synchronization")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/v1/sync/all",
        produces = { "application/json" }
    )
    
    default ResponseEntity<SyncResponse> syncAll(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"Synchronization started successfully\", \"status\" : \"ACCEPTED\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /api/v1/sync/expenses : Trigger synchronization of expenses only
     * Starts an asynchronous synchronization process for expenses only.
     *
     * @return Synchronization started successfully (status code 202)
     */
    @Operation(
        operationId = "syncExpenses",
        summary = "Trigger synchronization of expenses only",
        description = "Starts an asynchronous synchronization process for expenses only.",
        tags = { "Sync" },
        responses = {
            @ApiResponse(responseCode = "202", description = "Synchronization started successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SyncResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/v1/sync/expenses",
        produces = { "application/json" }
    )
    
    default ResponseEntity<SyncResponse> syncExpenses(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"Synchronization started successfully\", \"status\" : \"ACCEPTED\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /api/v1/sync/politicians : Trigger synchronization of politicians only
     * Starts an asynchronous synchronization process for politicians only.
     *
     * @return Synchronization started successfully (status code 202)
     */
    @Operation(
        operationId = "syncPoliticians",
        summary = "Trigger synchronization of politicians only",
        description = "Starts an asynchronous synchronization process for politicians only.",
        tags = { "Sync" },
        responses = {
            @ApiResponse(responseCode = "202", description = "Synchronization started successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SyncResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/v1/sync/politicians",
        produces = { "application/json" }
    )
    
    default ResponseEntity<SyncResponse> syncPoliticians(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"Synchronization started successfully\", \"status\" : \"ACCEPTED\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /api/v1/sync/presence : Trigger synchronization of presence only
     * Starts an asynchronous synchronization process for presence only.
     *
     * @return Synchronization started successfully (status code 202)
     */
    @Operation(
        operationId = "syncPresence",
        summary = "Trigger synchronization of presence only",
        description = "Starts an asynchronous synchronization process for presence only.",
        tags = { "Sync" },
        responses = {
            @ApiResponse(responseCode = "202", description = "Synchronization started successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SyncResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/v1/sync/presence",
        produces = { "application/json" }
    )
    
    default ResponseEntity<SyncResponse> syncPresence(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"Synchronization started successfully\", \"status\" : \"ACCEPTED\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /api/v1/sync/propositions : Trigger synchronization of propositions only
     * Starts an asynchronous synchronization process for propositions only.
     *
     * @return Synchronization started successfully (status code 202)
     */
    @Operation(
        operationId = "syncPropositions",
        summary = "Trigger synchronization of propositions only",
        description = "Starts an asynchronous synchronization process for propositions only.",
        tags = { "Sync" },
        responses = {
            @ApiResponse(responseCode = "202", description = "Synchronization started successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SyncResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/v1/sync/propositions",
        produces = { "application/json" }
    )
    
    default ResponseEntity<SyncResponse> syncPropositions(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"Synchronization started successfully\", \"status\" : \"ACCEPTED\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /api/v1/sync/speeches : Trigger synchronization of speeches only
     * Starts an asynchronous synchronization process for speeches only.
     *
     * @return Synchronization started successfully (status code 202)
     */
    @Operation(
        operationId = "syncSpeeches",
        summary = "Trigger synchronization of speeches only",
        description = "Starts an asynchronous synchronization process for speeches only.",
        tags = { "Sync" },
        responses = {
            @ApiResponse(responseCode = "202", description = "Synchronization started successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SyncResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/v1/sync/speeches",
        produces = { "application/json" }
    )
    
    default ResponseEntity<SyncResponse> syncSpeeches(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"Synchronization started successfully\", \"status\" : \"ACCEPTED\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /api/v1/sync/votes : Trigger synchronization of votes only
     * Starts an asynchronous synchronization process for votes only.
     *
     * @return Synchronization started successfully (status code 202)
     */
    @Operation(
        operationId = "syncVotes",
        summary = "Trigger synchronization of votes only",
        description = "Starts an asynchronous synchronization process for votes only.",
        tags = { "Sync" },
        responses = {
            @ApiResponse(responseCode = "202", description = "Synchronization started successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SyncResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/v1/sync/votes",
        produces = { "application/json" }
    )
    
    default ResponseEntity<SyncResponse> syncVotes(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"Synchronization started successfully\", \"status\" : \"ACCEPTED\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}